+++ MysticFixes/ErrorPolice.cs
@@ -25,17 +25,6 @@ namespace MiscFixes
     public class FixVanilla
     {
         /// <summary>
-        /// Halcyonite Shrine is able to drain 0 gold and will softlock itself.
-        /// If the scaled gold cost is less than 1, it gets truncated to 0 (mostly for modded scalings)
-        /// </summary>
-        [HarmonyPatch(typeof(HalcyoniteShrineInteractable), nameof(HalcyoniteShrineInteractable.Awake))]
-        [HarmonyPostfix]
-        public static void HalcyoniteShrineInteractable_Awake(HalcyoniteShrineInteractable __instance)
-        {
-            __instance.goldDrainValue = Math.Max(1, __instance.goldDrainValue);
-        }
-
-        /// <summary>
         /// Never checks the generic skill for null before overriding it, so ill assume all four of these should be handled the same
         /// RoR2.CharacterBody.HandleDisableAllSkillsDebuffg__HandleSkillDisableState|388_0 (System.Boolean _disable) (at:IL_0040)
         /// 
@@ -251,32 +240,6 @@ namespace MiscFixes
                 c.Emit(OpCodes.Brfalse, label);
             }
             else Log.PatchFail(il);
-        }
-
-        /// <summary>
-        /// Checks if the component exists, which should be the case for pre-sots overlay code
-        /// If true, it creates a temporary overlay instance from the component for backwards compatibility 
-        /// 
-        /// ArgumentNullException: Parameter name: source
-        /// 
-        /// UnityEngine.Material..ctor(UnityEngine.Material source) (at:IL_0008)
-        /// RoR2.TemporaryOverlayInstance.SetupMaterial() (at:IL_000D)
-        /// RoR2.TemporaryOverlayInstance.AddToCharacterModel(RoR2.CharacterModel characterModel) (at:IL_0000)
-        /// RoR2.TemporaryOverlay.AddToCharacerModel(RoR2.CharacterModel characterModel) (at:IL_0006)
-        /// 
-        /// UnityEngine.Material..ctor(UnityEngine.Material source) (at:IL_0008)
-        /// RoR2.TemporaryOverlayInstance.SetupMaterial() (at:IL_000D)
-        /// RoR2.TemporaryOverlayInstance.Start() (at:IL_0009)
-        /// RoR2.TemporaryOverlay.Start() (at:IL_0006)
-        /// </summary>
-        [HarmonyPatch(typeof(TemporaryOverlayInstance), nameof(TemporaryOverlayInstance.SetupMaterial))]
-        [HarmonyPrefix]
-        public static void TemporaryOverlayInstance_SetupMaterial(TemporaryOverlayInstance __instance)
-        {
-            if (!__instance.originalMaterial && __instance.componentReference && __instance.ValidateOverlay())
-            {
-                __instance.componentReference.CopyDataFromPrefabToInstance();
-            }
         }
 
         /// <summary>