+++ MysticFixes/SimpleFixes.cs
@@ -17,6 +17,65 @@ namespace MiscFixes
     public class SimpleFixes
     {
         /// <summary>
+        /// ConVars are not registered as lower case but when submitting them from the console they are converted, leading to a match fail.
+        /// </summary>
+        //[HarmonyPatch(typeof(Console), nameof(Console.RegisterConVarInternal))]
+        //[HarmonyILManipulator]
+        public static void FixConVarCaseSensitive(ILContext il)
+        {
+            var c = new ILCursor(il);
+            // Technically we aren't checking if ToLowerInvariant or ToLower(CultureInfo.InvariantCulture)
+            // is called after these instructions just like Console.Awake does for ConCommands, mostly
+            // because it's a hassle to check for the existence of either, but an extra call wouldn't hurt.
+            if (!c.TryGotoNext(
+                MoveType.After,
+                x => x.MatchLdarg(1),
+                x => x.MatchLdfld<RoR2.ConVar.BaseConVar>(nameof(RoR2.ConVar.BaseConVar.name))))
+            {
+                Log.PatchFail(il);
+                return;
+            }
+            c.Emit<string>(OpCodes.Callvirt, nameof(string.ToLowerInvariant));
+        }
+
+        /// <summary>
+        /// Checks if the component exists, which should be the case for pre-sots overlay code
+        /// If true, it creates a temporary overlay instance from the component for backwards compatibility 
+        /// 
+        /// ArgumentNullException: Parameter name: source
+        /// 
+        /// UnityEngine.Material..ctor(UnityEngine.Material source) (at:IL_0008)
+        /// RoR2.TemporaryOverlayInstance.SetupMaterial() (at:IL_000D)
+        /// RoR2.TemporaryOverlayInstance.AddToCharacterModel(RoR2.CharacterModel characterModel) (at:IL_0000)
+        /// RoR2.TemporaryOverlay.AddToCharacerModel(RoR2.CharacterModel characterModel) (at:IL_0006)
+        /// 
+        /// UnityEngine.Material..ctor(UnityEngine.Material source) (at:IL_0008)
+        /// RoR2.TemporaryOverlayInstance.SetupMaterial() (at:IL_000D)
+        /// RoR2.TemporaryOverlayInstance.Start() (at:IL_0009)
+        /// RoR2.TemporaryOverlay.Start() (at:IL_0006)
+        /// </summary>
+        [HarmonyPatch(typeof(TemporaryOverlayInstance), nameof(TemporaryOverlayInstance.SetupMaterial))]
+        [HarmonyPrefix]
+        public static void TemporaryOverlayInstance_SetupMaterial(TemporaryOverlayInstance __instance)
+        {
+            if (!__instance.originalMaterial && __instance.componentReference && __instance.ValidateOverlay())
+            {
+                __instance.componentReference.CopyDataFromPrefabToInstance();
+            }
+        }
+
+        /// <summary>
+        /// Halcyonite Shrine is able to drain 0 gold and will softlock itself.
+        /// If the scaled gold cost is less than 1, it gets truncated to 0 (mostly for modded scalings)
+        /// </summary>
+        [HarmonyPatch(typeof(HalcyoniteShrineInteractable), nameof(HalcyoniteShrineInteractable.Awake))]
+        [HarmonyPostfix]
+        public static void HalcyoniteShrineInteractable_Awake(HalcyoniteShrineInteractable __instance)
+        {
+            __instance.goldDrainValue = System.Math.Max(1, __instance.goldDrainValue);
+        }
+
+        /// <summary>
         /// Sometimes prevents loading, so suppress exceptions
         /// </summary>
         [HarmonyPatch(typeof(BaseSteamworks), nameof(BaseSteamworks.RunUpdateCallbacks))]
@@ -33,6 +92,7 @@ namespace MiscFixes
             if (!__instance.light)
                 __instance.enabled = false;
         }
+
         /// <summary>
         /// thanks to bubbet for the wonderful code
         /// 
@@ -85,28 +145,6 @@ namespace MiscFixes
                 });
             }
             else Log.PatchFail(il);
-        }
-
-        /// <summary>
-        /// ConVars are not registered as lower case but when submitting them from the console they are converted, leading to a match fail.
-        /// </summary>
-        [HarmonyPatch(typeof(Console), nameof(Console.RegisterConVarInternal))]
-        [HarmonyILManipulator]
-        public static void FixConVarCaseSensitive(ILContext il)
-        {
-            var c = new ILCursor(il);
-            // Technically we aren't checking if ToLowerInvariant or ToLower(CultureInfo.InvariantCulture)
-            // is called after these instructions just like Console.Awake does for ConCommands, mostly
-            // because it's a hassle to check for the existence of either, but an extra call wouldn't hurt.
-            if (!c.TryGotoNext(
-                MoveType.After,
-                x => x.MatchLdarg(1),
-                x => x.MatchLdfld<RoR2.ConVar.BaseConVar>(nameof(RoR2.ConVar.BaseConVar.name))))
-            {
-                Log.PatchFail(il);
-                return;
-            }
-            c.Emit<string>(OpCodes.Callvirt, nameof(string.ToLowerInvariant));
         }
     }
 }